+++ ether/__init__.py
@@ -3,4 +3,4 @@
 __title__ = "Æther"
 __author__ = "Atomic Junky"
 __licence__ = "GNU GPLv3"
-__version__ = "1.0.0rc5"
+__version__ = "1.0.0rc6"